requirements.txt
pwinput==1.0.2
pycryptodome==3.14.1
python_bcrypt==0.3.2
